AUDITFAIRNESS/requirements.txt
dash==2.17.1
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
evidently==0.4.33
python-dotenv==1.0.1
